move test util to firebase firestore pacakge
